+++ CHANGELOG.md
@@ -11,6 +11,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 - Added preliminary implementations for `FastX` views (within `irma-records` public API)
 - A `ValidatePaths` trait is now available in the library portion for ensuring no path is passed as both an input and an output, and that all output paths are distinct
+- `WriteRecord` is now compatible with `FastX`
 
 ### Changed